+++ src/components/home/components/MCQ_view.tsx
