+++ src/components/home/components/featured_image_gallery.tsx
