+++ src/components/home/components/mcq/reader/parts/hint_pane.tsx
