+++ src/components/home/components/mcq/shared/index.ts
