+++ src/components/home/components/navigatePlayerButton/index.tsx
@@ -11,7 +11,7 @@ const NavigatePlayerButton = ({ isDisplay, src }: DisplayMusicPlayerButtonProps)
         // TODO change domain URL not use localhost
         const windowFeatures = 'width=412,height=732,top=100,left=100'; // 412 × 732
         const encodedMessage = encodeURIComponent(src); // 쿼리 파라미터로 안전하게 전달
-        const url = `http://localhost:3000/watch?message=${encodedMessage}`;
+        const url = `http://localhost:4000/watch?message=${encodedMessage}`;
 
         window.open(url, '_blank', windowFeatures);
     }, [src]); // `src`를 의존성 배열에 추가
